+++ soxs/tests/test_power_law.py
@@ -4,9 +4,10 @@ import shutil
 import tempfile
 from soxs.spectra import Spectrum, get_tbabs_absorb
 from soxs.spatial import PointSourceModel
-from soxs.simput import PhotonList, SimputCatalog
+from soxs.simput import SimputCatalog
 from soxs.instrument_registry import \
-    get_instrument_from_registry
+    get_instrument_from_registry, \
+    make_simple_instrument
 from soxs.instrument import instrument_simulator, \
     RedistributionMatrixFile, AuxiliaryResponseFile
 from soxs.utils import convert_rmf
@@ -15,6 +16,8 @@ from sherpa.astro.ui import load_user_model, add_user_pars, \
     load_pha, ignore, fit, set_model, set_stat, set_method, \
     covar, get_covar_results, set_covar_opt
 from numpy.random import RandomState
+
+make_simple_instrument("hdxi", "new_hdxi", 20.0, 1024)
 
 prng = RandomState(69)
 
@@ -41,7 +44,7 @@ def plaw_fit(alpha_sim):
 
     exp_time = (50.0, "ks")
     area = 40000.0
-    inst_name = "hdxi"
+    inst_name = "new_hdxi"
 
     spec = Spectrum.from_powerlaw(alpha_sim, redshift, norm_sim, 0.1, 10.0, 20000)
     spec.apply_foreground_absorption(nH_sim, model="tbabs")